refactoring api again
